+++ MSGraph/functions/exchange/Move-MgaMailMessage.ps1
@@ -38,7 +38,7 @@
 
         Moves messages in variable $mails to the folder in the variable $destinationFolder.
         also possible:
-        PS C:\> Move-MgaMailMessage -InputObject $mails -DestinationFolder $destinationFolder
+        PS C:\> Move-MgaMailMessage -Message $mails -DestinationFolder $destinationFolder
 
         The variable $mails can be represent:
         PS C:\> $mails = Get-MgaMailMessage -Folder Inbox -ResultSize 1
@@ -50,7 +50,7 @@
         PS C:\> Move-MgaMailMessage -Id $mails.id -DestinationFolder $destinationFolder
 
         Moves messages into the folder $destinationFolder.
-        
+
         The variable $mails can be represent:
         PS C:\> $mails = Get-MgaMailMessage -Folder Inbox -ResultSize 1
 
@@ -65,27 +65,21 @@
         PS C:\> $destinationFolder = Get-MgaMailFolder -Name "Archive"
 
     #>
-    [CmdletBinding(SupportsShouldProcess = $true, ConfirmImpact = 'Medium', DefaultParameterSetName = 'ByInputObject')]
+    [CmdletBinding(SupportsShouldProcess = $true, ConfirmImpact = 'Medium', DefaultParameterSetName = 'Default')]
     [Alias()]
     [OutputType([MSGraph.Exchange.Mail.Message])]
     param (
-        [Parameter(Mandatory = $true, ValueFromPipeline = $true, ValueFromPipelineByPropertyName = $true, ParameterSetName = 'ByInputObject')]
-        [Alias("Message")]
-        [MSGraph.Exchange.Mail.Message]
-        $InputObject,
+        [Parameter(Mandatory=$true, Position = 0, ValueFromPipeline = $true, ValueFromPipelineByPropertyName = $true)]
+        [Alias('InputObject', 'MessageId', 'Id', 'Mail', 'MailId')]
+        [MSGraph.Exchange.Mail.MessageParameter[]]
+        $Message,
 
-        [Parameter(Mandatory = $true, ValueFromPipeline = $true, ValueFromPipelineByPropertyName = $true, ParameterSetName = 'ById')]
-        [Alias("MessageId")]
-        [string[]]
-        $Id,
+        [Parameter(Mandatory = $true, Position = 1)]
+        [MSGraph.Exchange.Mail.FolderParameter]
+        $DestinationFolder,
 
-        [Parameter(ValueFromPipeline = $true, ValueFromPipelineByPropertyName = $true, ParameterSetName = 'ById')]
         [string]
         $User,
-
-        [Parameter(Mandatory = $true)]
-        [MSGraph.Exchange.Mail.Folder]
-        $DestinationFolder,
 
         [MSGraph.Core.AzureAccessToken]
         $Token,
@@ -94,27 +88,17 @@
         $PassThru
     )
     begin {
-    }
-
-    process {
-        $messages = @()
-
-        # Get input from pipeable objects
-        Write-PSFMessage -Level Debug -Message "Gettings messages by parameter set $($PSCmdlet.ParameterSetName)" -Tag "ParameterSetHandling"
-        switch ($PSCmdlet.ParameterSetName) {
-            "ByInputObject" {
-                $messages = $InputObject.Id
-                $User = $InputObject.BaseObject.User
-            }
-            "ById" {
-                $messages = $Id
-            }
-            Default { Stop-PSFFunction -Tag "ParameterSetHandling" -Message "Unhandled parameter set. ($($PSCmdlet.ParameterSetName)) Developer mistage." -EnableException $true -Exception ([System.Management.Automation.RuntimeException]::new("Unhandled parameter set. ($($PSCmdlet.ParameterSetName)) Developer mistage.")) -FunctionName $MyInvocation.MyCommand }
+        if($DestinationFolder.TypeName -like "System.String") {
+            [MSGraph.Exchange.Mail.FolderParameter]$DestinationFolder = Get-MgaMailFolder -Name $DestinationFolder.ToString() -User $User -Token $Token
         }
 
         $bodyHash = @{
             destinationId = ($DestinationFolder.Id | ConvertTo-Json)
         }
+    }
+
+    process {
+        Write-PSFMessage -Level Debug -Message "Gettings messages by parameter set $($PSCmdlet.ParameterSetName)" -Tag "ParameterSetHandling"
 
         #region Put parameters (JSON Parts) into a valid "message"-JSON-object together
         $bodyJsonParts = @()
@@ -125,11 +109,29 @@
         #endregion Put parameters (JSON Parts) into a valid "message"-JSON-object together
 
         #region move messages
-        foreach ($messageId in $messages) {
-            if ($pscmdlet.ShouldProcess("messageId $($messageId)", "Move to folder '$($DestinationFolder.Name)'")) {
-                Write-PSFMessage -Tag "MessageUpdate" -Level Verbose -Message "Move messageId '$($messageId)' to folder '$($DestinationFolder.Name)'"
+        foreach ($messageItem in $Message) {
+            if ($messageItem.TypeName -like "System.String") {
+                if ($messageItem.Id -and ($messageItem.Id.Length -eq 152)) {
+                    [MSGraph.Exchange.Mail.MessageParameter]$messageItem = Get-MgaMailMessage -InputObject $messageItem.Id -User $User -Token $Token
+                }
+                else {
+                    Write-PSFMessage -Level Warning -Message "The specified input string seams not to be a valid Id. Skipping object '$($messageItem)'" -Tag "InputValidation"
+                    continue
+                }
+            }
+
+            if ($User -and ($messageItem.TypeName -like "MSGraph.Exchange.Mail.Message") -and ($User -notlike $messageItem.InputObject.BaseObject.User)) {
+                Write-PSFMessage -Level Important -Message "Individual user specified with message object! User from message object ($($messageItem.InputObject.BaseObject.User))will take precedence on specified user ($($User))!" -Tag "InputValidation"
+                $User = $messageItem.InputObject.BaseObject.User
+            }
+            elseif ((-not $User) -and ($messageItem.TypeName -like "MSGraph.Exchange.Mail.Message")) {
+                $User = $messageItem.InputObject.BaseObject.User
+            }
+
+            if ($pscmdlet.ShouldProcess("message '$($messageItem)'", "Move to folder '$($DestinationFolder.Name)'")) {
+                Write-PSFMessage -Tag "MessageUpdate" -Level Verbose -Message "Move message '$($messageItem)' to folder '$($DestinationFolder)'"
                 $invokeParam = @{
-                    "Field"        = "messages/$($messageId)/move"
+                    "Field"        = "messages/$($messageItem.Id)/move"
                     "User"         = $User
                     "Body"         = $bodyJSON
                     "ContentType"  = "application/json"
@@ -137,9 +139,7 @@
                     "FunctionName" = $MyInvocation.MyCommand
                 }
                 $output = Invoke-MgaPostMethod @invokeParam
-                if ($PassThru) {
-                    [MSGraph.Exchange.Mail.Message]@{ BaseObject = $output }
-                }
+                if ($PassThru) { New-MgaMailMessageObject -RestData $output }
             }
         }
         #endregion Update messages